Refactor store_test.py
